requirements.txt
discord.py
riotwatcher>=3.2.1
python-dotenv
aiohttp
